+++ examples/admission_controller.rs
@@ -12,31 +12,25 @@ use warp::{reply, Filter, Reply};
 
 #[tokio::main]
 async fn main() {
-    std::env::set_var("RUST_LOG", "debug,kube=debug");
-    tracing_subscriber::fmt::init();
+    std::env::set_var("RUST_LOG", "info,kube=debug");
+    env_logger::init();
 
     let routes = warp::path("mutate")
         .and(warp::body::json())
         .and_then(mutate_handler)
         .with(warp::trace::request());
 
-
-    // For admission controllers running behind a service, you must generate a
-    // certificate for the service and encode the CA in the
-    // MutatingWebhookConfiguration, then terminate TLS here. For controllers
-    // running outside of the cluster, while still necessary, it's fine to
-    // handle TLS upstream. See:
+    // You must generate a certificate for the service / url,
+    // encode the CA in the MutatingWebhookConfiguration, and terminate TLS here.
+    // See admission_setup.sh + admission_controller.yaml.tpl
     // https://kubernetes.io/blog/2019/03/21/a-guide-to-kubernetes-admission-controllers/#tls-certificates
-
-    // warp::serve(warp::post().and(routes))
-    //     .tls()
-    //     .cert_path("admission-controller-tls.crt")
-    //     .key_path("admission-controller-tls.key")
-    //     .run(([0, 0, 0, 0], 8443))
-    //     .await;
-
+    let addr = format!("{}:8443", std::env::var("ADMISSION_PRIVATE_IP").unwrap());
     warp::serve(warp::post().and(routes))
-        .run(([0, 0, 0, 0], 8080))
+        .tls()
+        .cert_path("admission-controller-tls.crt")
+        .key_path("admission-controller-tls.key")
+        //.run(([0, 0, 0, 0], 8443))
+        .run(addr.parse::<std::net::SocketAddr>().unwrap())
         .await;
 }
 
@@ -69,34 +63,42 @@ async fn mutate_handler(body: AdmissionReview<DynamicObject>) -> Result<impl Rep
             }
         };
     };
-
+    // Need to return an AdmissionResponse wrapped in an AdmissionReview
     Ok(reply::json(&res.into_review()))
 }
 
 fn mutate(
     mut res: AdmissionResponse,
     obj: DynamicObject,
-    dry_run: bool,
+    _dry_run: bool, // no external operations performed, no need to track dry_run
 ) -> Result<AdmissionResponse, Box<dyn Error>> {
-    let mut patches: Vec<json_patch::PatchOperation> = Vec::new();
     let labels = obj.metadata.labels.unwrap_or_default();
 
-    // If the resource contains a label named "test_reject", it will be
-    // forbidden from creating.
-    if labels.contains_key("test_reject") {
-        return Err("Resource contained 'test_reject' label".into());
+    // If the resource contains an "illegal" label, we reject it
+    if labels.contains_key("illegal") {
+        return Err("Resource contained 'illegal' label".into());
     }
 
-    // If the resource doesn't contain "test_modify", we add it to the
-    // resource.
-    if !labels.contains_key("test_modify") {
-        patches.push(json_patch::PatchOperation::Add(json_patch::AddOperation {
-            path: "/metadata/labels/test_modify".to_owned(),
-            value: serde_json::Value::String("modified-by-admission-controller".to_owned()),
-        }))
-    }
+    // If the resource doesn't contain "admission", we add it to the resource.
+    let patches = if !labels.contains_key("admission") {
+        use json_patch::{AddOperation, PatchOperation};
+        vec![
+            // Ensure labels exist before adding a key to it
+            PatchOperation::Add(AddOperation {
+                path: "/metadata/labels".to_string(),
+                value: serde_json::json!({}),
+            }),
+            // Add our label
+            PatchOperation::Add(AddOperation {
+                path: "/metadata/labels/admission".to_owned(),
+                value: serde_json::Value::String("modified-by-admission-controller".to_owned()),
+            }),
+        ]
+    } else {
+        vec![]
+    };
 
-    if !dry_run && !patches.is_empty() {
+    if !patches.is_empty() {
         res = res.with_patch(json_patch::Patch(patches))?;
     }